requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
openpyxl==3.1.2
pandas>=2.1.0
numpy>=1.26.0
PyPDF2==3.0.1